+++ app/src/main/java/nextstep/payments/MainActivity.kt
@@ -3,14 +3,8 @@ package nextstep.payments
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import nextstep.payments.ui.theme.PaymentsTheme
+import nextstep.payments.ui.newcard.NewCardScreen
+import nextstep.payments.designsystem.theme.PaymentsTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
