+++ app/src/main/java/nextstep/payments/ui/newcard/NewCardScreen.kt
@@ -1,24 +1,26 @@
-package nextstep.payments
+package nextstep.payments.ui.newcard
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import nextstep.payments.ui.theme.PaymentsTheme
+import nextstep.payments.designsystem.component.PaymentCard
+import nextstep.payments.designsystem.theme.PaymentsTheme
+import nextstep.payments.ui.newcard.component.CardExpiredDateTextFiled
+import nextstep.payments.ui.newcard.component.CardNumberTextFiled
+import nextstep.payments.ui.newcard.component.CardOwnerNameTextFiled
+import nextstep.payments.ui.newcard.component.CardPasswordTextFiled
+import nextstep.payments.ui.newcard.component.NewCardTopBar
 
 @Composable
 fun NewCardScreen(
@@ -39,6 +41,7 @@ fun NewCardScreen(
         setExpiredDate = viewModel::setExpiredDate,
         setOwnerName = viewModel::setOwnerName,
         setPassword = viewModel::setPassword,
+        modifier = modifier
     )
 }
 
@@ -56,7 +59,7 @@ fun NewCardScreen(
     modifier: Modifier = Modifier,
 ) {
     Scaffold(
-        topBar = { NewCardTopBar(onBackClick = { TODO() }, onSaveClick = { TODO() }) },
+        topBar = { NewCardTopBar(onBackClick = { }, onSaveClick = { }) },
         modifier = modifier
     ) { innerPadding ->
         Column(
@@ -72,38 +75,10 @@ fun NewCardScreen(
 
             Spacer(modifier = Modifier.height(10.dp))
 
-            OutlinedTextField(
-                value = cardNumber,
-                onValueChange = setCardNumber,
-                label = { Text("카드 번호") },
-                placeholder = { Text("0000 - 0000 - 0000 - 0000") },
-                modifier = Modifier.fillMaxWidth(),
-            )
-
-            OutlinedTextField(
-                value = expiredDate,
-                onValueChange = setExpiredDate,
-                label = { Text("만료일") },
-                placeholder = { Text("MM / YY") },
-                modifier = Modifier.fillMaxWidth(),
-            )
-
-            OutlinedTextField(
-                value = ownerName,
-                onValueChange = setOwnerName,
-                label = { Text("카드 소유자 이름(선택)") },
-                placeholder = { Text("카드에 표시된 이름을 입력하세요.") },
-                modifier = Modifier.fillMaxWidth(),
-            )
-
-            OutlinedTextField(
-                value = password,
-                onValueChange = setPassword,
-                label = { Text("비밀번호") },
-                placeholder = { Text("0000") },
-                modifier = Modifier.fillMaxWidth(),
-                visualTransformation = PasswordVisualTransformation(),
-            )
+            CardNumberTextFiled(value = cardNumber, onValueChange = setCardNumber)
+            CardExpiredDateTextFiled(value = expiredDate, onValueChange = setExpiredDate)
+            CardOwnerNameTextFiled(value = ownerName, onValueChange = setOwnerName)
+            CardPasswordTextFiled(value = password, onValueChange = setPassword)
         }
     }
 }
@@ -113,8 +88,8 @@ fun NewCardScreen(
 private fun StatefulNewCardScreenPreview() {
     PaymentsTheme {
         NewCardScreen(viewModel = NewCardViewModel().apply {
-            setCardNumber("0000 - 0000 - 0000 - 0000")
-            setExpiredDate("00 / 00")
+            setCardNumber("0000000000000000")
+            setExpiredDate("0000")
             setOwnerName("홍길동")
             setPassword("0000")
         })
@@ -126,8 +101,8 @@ private fun StatefulNewCardScreenPreview() {
 private fun StatelessNewCardScreenPreview() {
     PaymentsTheme {
         NewCardScreen(
-            cardNumber = "0000 - 0000 - 0000 - 0000",
-            expiredDate = "00 / 00",
+            cardNumber = "0000000000000000",
+            expiredDate = "0000",
             ownerName = "홍길동",
             password = "0000",
             setCardNumber = {},
